Fix float bug in loops of texture code

+++ src/trainTextureTransfer.py
@@ -43,7 +43,7 @@ def main():
     X = to_var(X) if opts.gpu else X
     Noise = to_var(Noise) if opts.gpu else Noise
     for epoch in range(opts.texture_step1_epochs):
-        for i in range(opts.Ttraining_num/opts.batchsize):
+        for i in range(int(opts.Ttraining_num/opts.batchsize)):
             x, y = cropping_training_batches(X, Y, Noise, opts.batchsize, 
                                       opts.Tanglejitter, opts.subimg_size, opts.subimg_size)
             losses = netShapeM.texture_one_pass(x, y)
